refactor: 项目重构 — 组件拆分、类型安全、代码模块化
- **Graph Visualization**: 提取 types.ts, useGraphData.ts, Tooltip.tsx, ClearConfirmDialog.tsx，移除 any 类型
- **State Management**: 重构事件总线为闭包架构
- **Converter Tools**: 提取 utils.ts 工具模块
- **UI Components**: 清理未使用的代码、修复类型定义
- **Lint Fixes**: 修复 setState in effect、any 类型等问题
