+++ docs/superpowers/specs/2026-04-12-recommended-path-cards-design.md
@@ -1,0 +1,75 @@
+# 推荐链路卡片 — 设计文档
+
+## 背景
+
+当前推荐链路功能：用户在图谱点击节点后，RightPanel 底部显示推荐笔记列表（标题 + 分数 + 领域色条）。用户无法直观了解推荐理由，且需要切换注意力到右侧面板。
+
+## 目标
+
+在图谱上直接以卡片形式展示 Top 3 推荐笔记的预览信息，提升信息获取效率，让用户无需离开图谱上下文就能决定下一步。
+
+## 交互流程
+
+1. **点击图谱任意节点** → 系统计算推荐节点（B、C、D）
+2. **Top 3 推荐卡自动展开** → 各自显示在被推荐节点的旁边
+3. **点击任意卡片** → 进入该笔记（selectNode）
+4. **点击图谱空白处 / 切换节点** → 所有卡片关闭
+5. **RightPanel 底部推荐列表** → 保留，作为补充参考
+
+## 卡片设计
+
+### 内容结构
+
+```
+┌─────────────────────────────────┐
+│ ● Agent记忆架构设计      87%   │  ← 领域色点 · 标题 · 右上分数
+│   AI 智能体与工程 · 链接笔记    │  ← 领域 · 类型（次要信息）
+├─────────────────────────────────┤
+│ 本文系统梳理了 Agent 记忆架构的  │  ← 部分正文摘要（3 行）
+│ 几种主流方案，包括 Senser-       │    maxHeight: 72px
+│ Remember-Recall 框架……          │
+└─────────────────────────────────┘
+```
+
+- 无 AI 理由文本
+- 无按钮（点击卡片本身进入笔记）
+- 无关闭按钮（通过点击空白处或切换节点关闭）
+
+### 卡片位置
+
+跟随节点位置——浮动在被推荐节点的附近（由图谱节点坐标决定偏移方向，避免遮挡节点本身）。
+
+**不允许卡片重叠**——如果多个推荐节点距离过近导致卡片会互相遮挡，则对这些卡片进行位置错开（如交替向左/右偏移，或上下排列）。
+
+### 视觉样式
+
+- 宽度：220–240px
+- 圆角：10px
+- 背景：白色，1px 边框 + 柔和阴影
+- 标题：12px，weight 600，单行省略
+- 分数：右上角，13px，weight 700，颜色 = 领域色
+- 正文摘要：11px，maxHeight 72px（4 行），lineClamp
+
+### 唯一性
+
+同一时刻只存在一套推荐卡片（对应当前选中节点），切换选中节点时旧卡片关闭，新卡片打开。
+
+## 技术要点
+
+- 推荐节点通过 `recommendedPaths`（store 中已有）获取，取前 3 个
+- 卡片跟随节点位置：读取 ForceGraph 内部节点坐标，在图谱 DOM 坐标系中定位卡片（需考虑 `fgRef` 获取节点位置）
+- 关闭机制：监听图谱 canvas 的 `click` 事件，判断是否命中节点 or 卡片，否则关闭
+- 卡片与节点点击事件需要阻止冒泡，避免误触发关闭
+
+## 数据来源
+
+- `recommendedPaths[n].title`
+- `recommendedPaths[n].compositeScore` → 百分比
+- `recommendedPaths[n].bodyPreview` 或 `note.body` 截取前 N 字符
+- `recommendedPaths[n].domain` → `DOMAIN_COLORS[domain]`
+
+## 非功能
+
+- 不修改 `recommend.ts` 算法
+- 不修改 `graphStore` 状态（只读）
+- RightPanel footer 保持不变
